+++ CHANGELOG.md
@@ -12,5 +12,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added tests ([bc9af56](ssh://git@git:7999/~blaudiy/talismans/commit/bc9af565640370e954ff3db4e6998c8e8e241920) by Yann BLAUDIN DE THE).
 - Added CHANGELOG ([1560b2c](ssh://git@git:7999/~blaudiy/talismans/commit/1560b2c051c6562b9059dbb0424d166641fbf9e8) by Yann BLAUDIN DE THE).